Disable JSON content-type validation
A regression from converting from requests to aiohttp, which requires JSON mandated content-type.
HT-ZF9 is not sending the mandatory content-type header for its guide endpoint,
causing it to fail to ContentTypeError. As we do not really care about the correctness,
this commit disables the content type checking as described in https://docs.aiohttp.org/en/stable/client_advanced.html#disabling-content-type-validation-for-json-responses

Fixes #58