Allow forcing getSupportedApiInfo response from local devinfo file
Some devices do not support the ability to obtain information about available APIs.
This PR adds new --devinfo-file option that allows using an existing devinfo file to feed the list of available services,
which may be helpful for such devices as long as they implement getMethodTypes() call for the services.

Potential fix for #29