Add a NotificationCallback type annotation.
This allows typing the callback information and ensuring the passed
callback functions are ready to expect what they would receive.

This may be a bit more complicated, as callbacks are going to receive
specific notifications rather than the most generic one of them, but I
think this is the better option.