Add support for versioned methods used by newer devices (#132)
This updates to allow the support of multiple API method versions returned by `getMethodTypes`. It will also check if the version currently being used is a specifically supported version (`getSupportedApiInfo`) for the device and log a debug message if not.

Adds support for the following specific API Methods
- version `1.2` for `getCurrentExternalTerminalsStatus`. Includes what should be a non-breaking change for `notifyPlayingContentInfo` notification.
- version `1.3` for `getSourceList`.

Update method parameters for the following, as newer devices can be stricter with the API spec. Should be non-breaking.
- `getCustomEqualizerSettings`.
- `getSpeakerSettings`.

Adds support for Sony STR-AN1000.
Limited backward compatibility testing has been done with a STR-DN1080 and a HT-XT3.